+++ src/lib/constants.sh
@@ -16,5 +16,4 @@ readonly VALID_SUBCOMMANDS=(
   config
   apply
   uninstall
-  gpipe
 )